# server/docs/swagger.yaml
basePath: /api
definitions:
  handlers.CreateUserRequest:
    properties:
      clan_group_id:
        type: integer
      group_id:
        type: integer
      name:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - clan_group_id
    - group_id
    - name
    - password
    - username
    type: object
  handlers.LoginRequest:
    properties:
      password:
        example: password
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  models.AlbumResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.IdentityListResponse:
    properties:
      identities:
        items:
          $ref: '#/definitions/models.IdentityResponse'
        type: array
    type: object
  models.IdentityResponse:
    properties:
      created_at:
        type: string
      group_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.UpdateIdentityByIdAndGroupIdRequest:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  services.MediaItemRange:
    properties:
      month:
        type: integer
      year:
        type: integer
    type: object
info:
  contact: {}
paths:
  /album/write:
    get:
      description: Get albums for write
      parameters:
      - description: Authorization
        example: bearer token
        format: bearer
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AlbumResponse'
            type: array
      tags:
      - Album
  /auth/login:
    post:
      consumes:
      - application/json
      description: login
      parameters:
      - description: Login credentials
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /identities:
    get:
      description: identity 목록을 조회
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdentityListResponse'
      tags:
      - Identity
  /identity:
    put:
      description: identity를 수정
      parameters:
      - description: Update Identity Request
        in: body
        name: updateIdentityRequest
        required: true
        schema:
          $ref: '#/definitions/models.UpdateIdentityByIdAndGroupIdRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdentityResponse'
      tags:
      - Identity
  /identity/{id}:
    get:
      description: identity를 조회
      parameters:
      - description: Identity ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdentityResponse'
      tags:
      - Identity
  /media-item:
    get:
      description: 미디어 아이템 목록 조회
      parameters:
      - description: Authorization
        example: bearer token
        format: bearer
        in: header
        name: Authorization
        required: true
        type: string
      - description: From
        example: "2025-01-01"
        in: query
        name: from
        required: true
        type: string
      - description: To
        example: "2025-01-01"
        in: query
        name: to
        required: true
        type: string
      responses:
        "200":
          description: OK
      tags:
      - MediaItem
  /media-item/image/upload:
    post:
      consumes:
      - multipart/form-data
      description: 이미지 업로드
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      - description: Album ID
        in: query
        name: album_id
        required: true
        type: string
      - description: Upload Batch ID
        in: query
        name: upload_batch_id
        required: true
        type: string
      - description: Retry
        example: "1"
        in: query
        name: retry
        type: string
      - description: Authorization
        example: bearer token
        format: bearer
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      tags:
      - MediaItem
  /media-item/range:
    get:
      description: 사진이 있는 년도와 월 목록 조회
      parameters:
      - description: Authorization
        example: bearer token
        format: bearer
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/services.MediaItemRange'
            type: array
      tags:
      - MediaItem
  /media-item/upload-batch:
    post:
      description: 업로드 배치 생성
      parameters:
      - description: Album ID
        in: query
        name: album_id
        required: true
        type: string
      - description: Authorization
        example: bearer token
        format: bearer
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      tags:
      - MediaItem
  /media-item/upload-batch/status:
    get:
      description: 업로드 배치 상태 조회
      parameters:
      - description: Upload Batch ID
        in: query
        name: upload_batch_id
        required: true
        type: string
      - description: Authorization
        example: bearer token
        format: bearer
        in: header
        name: Authorization
        required: true
        type: string
      responses: {}
      tags:
      - MediaItem
  /user:
    post:
      description: create user
      parameters:
      - description: Create User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      tags:
      - User
swagger: "2.0"