# server/internal/api/handlers/mediaItemHandler.go
package handlers

import (
	"log"
	"time"

	"github.com/labstack/echo/v4"

	"github.com/ywl0806/yuno_kiroku/internal/api/handlers/models"
	"github.com/ywl0806/yuno_kiroku/internal/api/middlewares"
	"github.com/ywl0806/yuno_kiroku/internal/api/services"
	"github.com/ywl0806/yuno_kiroku/internal/api/utils"
	"github.com/ywl0806/yuno_kiroku/internal/db"

	"github.com/ywl0806/yuno_kiroku/internal/api/appErrors"
)

type MediaItemHandler struct {
	mediaItemService *services.MediaItemService
	faceService      *services.FaceService
}

func NewMediaItemHandler(
	mediaItemService *services.MediaItemService,
	faceService *services.FaceService,
) *MediaItemHandler {
	return &MediaItemHandler{
		mediaItemService: mediaItemService,
		faceService:      faceService,
	}
}

// @Tags MediaItem
// @Description 업로드 배치 생성
// @Param album_id query string true "Album ID"
// @Param Authorization header string true "Authorization" format(bearer) example(bearer token)
// @Router /media-item/upload-batch [post]
func (con *MediaItemHandler) CreateUploadBatch(c echo.Context) error {
	// album 쿼리 파라미터 가져오기
	albumId, err := utils.ConvertToInt32(c.QueryParam("album_id"))
	if err != nil {
		return appErrors.NewValidationError(utils.GetMessage("", map[string]string{"field": "album_id"}))
	}

	// 그룹 ID 가져오기
	authUser := middlewares.GetAuthUser(c)
	groupId := authUser.GroupId

	// 컨텍스트 가져오기
	ctx := c.Request().Context()

	// 업로드 배치 생성
	uploadBatch, err := con.mediaItemService.CreateUploadBatch(ctx, groupId, albumId)
	if err != nil {
		return err
	}

	return c.JSON(200, map[string]interface{}{
		"id":         uploadBatch.ID,
		"album_id":   uploadBatch.AlbumID,
		"created_at": uploadBatch.CreatedAt,
	})
}

// @Tags MediaItem
// @Description 이미지 업로드
// @Accept  multipart/form-data
// @Param file formData file true "file"
// @Param album_id query string true "Album ID"
// @Param upload_batch_id query string true "Upload Batch ID"
// @Param retry query string false "Retry" example(1)
// @Param Authorization header string true "Authorization" format(bearer) example(bearer token)
// @Router /media-item/image/upload [post]
func (con *MediaItemHandler) UploadImage(c echo.Context) error {
	file, err := c.FormFile("file")
	if err != nil {
		return appErrors.NewValidationError("")
	}

	albumId, err := utils.ConvertToInt32(c.QueryParam("album_id"))
	if err != nil {
		return appErrors.NewValidationError("")
	}

	uploadBatchID, err := utils.ConvertToInt32(c.QueryParam("upload_batch_id"))
	if err != nil {
		return appErrors.NewValidationError("")
	}

	retry, _ := utils.ConvertToBool(c.QueryParam("retry"))

	authUser := middlewares.GetAuthUser(c)
	groupId := authUser.GroupId
	ctx := c.Request().Context()

	result, err := con.mediaItemService.UploadImage(ctx, file, groupId, albumId, uploadBatchID, retry)
	if err != nil {
		log.Println("upload image error: ", err)
		return err
	}

	return c.JSON(200, map[string]interface{}{
		"media_item_id": result.MediaItemID,
		"status":        result.Status,
	})
}

// @Tags MediaItem
// @Description 사진이 있는 년도와 월 목록 조회
// @Router /media-item/range [get]
// @Param Authorization header string true "Authorization" format(bearer) example(bearer token)
// @Success 200 {object} []services.MediaItemRange
func (con *MediaItemHandler) GetMediaItemRange(c echo.Context) error {

	authUser := middlewares.GetAuthUser(c)
	ranges, err := con.mediaItemService.GetMediaItemRange(c.Request().Context(), authUser.ClanGroupId)
	if err != nil {
		return err
	}
	return c.JSON(200, ranges)
}

type GetMediaItemsRequest struct {
	From *time.Time `query:"from" validate:"required"`
	To   *time.Time `query:"to" validate:"required"`
}

func (GetMediaItemsRequest) bind(c echo.Context, p *db.GetMediaItemsByTakenAtParams) error {
	reqParams := new(GetMediaItemsRequest)
	if err := c.Bind(reqParams); err != nil {
		return err
	}
	if err := c.Validate(reqParams); err != nil {
		return err
	}

	authUser := middlewares.GetAuthUser(c)

	p.TakenAtFrom = *reqParams.From
	p.TakenAtTo = *reqParams.To
	p.ClanGroupID = authUser.ClanGroupId
	return nil
}

// @Tags MediaItem
// @Description 미디어 아이템 목록 조회
// @Router /media-item [get]
// @Param Authorization header string true "Authorization" format(bearer) example(bearer token)
// @Param from query string true "From" example(2025-01-01)
// @Param to query string true "To" example(2025-01-01)
// @Success 200
func (con *MediaItemHandler) GetMediaItems(c echo.Context) error {
	reqParams := new(GetMediaItemsRequest)

	params := new(db.GetMediaItemsByTakenAtParams)

	if err := reqParams.bind(c, params); err != nil {
		log.Println("bind get media items request error: ", err)
		return err
	}
	mediaItems, err := con.mediaItemService.GetMediaItemsByTakenAt(c.Request().Context(), params)
	if err != nil {
		log.Println("get media items by taken at error: ", err)
		return err
	}

	return c.JSON(200, models.NewMediaItemsResponse(mediaItems))

}

// @Tags MediaItem
// @Description 업로드 배치 상태 조회
// @Param upload_batch_id query string true "Upload Batch ID"
// @Param Authorization header string true "Authorization" format(bearer) example(bearer token)
// @Router /media-item/upload-batch/status [get]
func (con *MediaItemHandler) GetUploadBatchStatus(c echo.Context) error {
	// upload_batch_id 쿼리 파라미터 가져오기
	uploadBatchID, err := utils.ConvertToInt32(c.QueryParam("upload_batch_id"))
	if err != nil {
		return appErrors.NewValidationError("")
	}

	// 컨텍스트 가져오기
	ctx := c.Request().Context()

	// 업로드 배치 조회
	uploadStatuses, err := con.mediaItemService.GetUploadStatuses(ctx, uploadBatchID)
	if err != nil {
		return err
	}

	uploadStatusesResponse := models.NewUploadBatchStatusResponse(uploadStatuses)

	return c.JSON(200, uploadStatusesResponse)
}